+++ pytest_automock/__init__.py
@@ -3,5 +3,5 @@
 from .mock import *
 
 
-__version__ = "0.3.0"
+__version__ = "0.4.0"
 version = tuple(map(int, __version__.split(".")))